[mock-chronik-client] Update syntax to match chronik-client 0.9.0
Summary:
Update mock-chronik-client to match latest syntax. Add a unit test to confirm `ws.subs` is now supported.

Note: ChronikClient no longer has `_url` and `_wsUrl` properties. These are removed. These were not used in tests, but were included as a way of referencing that this was a mock chronik client and not an actual, connected chronik client. imo not really a value add, but now also incorrect, so has to go.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14871