[chronik-client] Apply standardized-ish commenting
Summary:
Use single line comments `//` for comments of 1 or 2 lines. Use multi-line commments `/* */` for longer comments, not including comments on the first and last line of multi-line block

This is the kind of thing that should be linted but I was not able to find support for this in either `eslint` or `prettier`. Because this issue has not come up in any other repo, just patching it here and moving on is best available approach.

Test Plan: Confirm changes are only to comments

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14884