[secp256k1] remove VERIFY_SETUP define
Summary:
This define was seemingly introduced for VERIFY mode code with side
effects (for setup purposes), that should just be executed without any
checks. The same can be achieved by putting it in an `#if VERIFY` block,
so we can remove it.

This concludes backport of [[https://github.com/bitcoin-core/secp256k1/pull/1393 | secp256k1#1393]]
https://github.com/bitcoin-core/secp256k1/pull/1393/changes/bb4672342efce7fae1cfd30e007c6835a25286a7

Depends on D20169

Test Plan: `ninja check-secp256k1`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20170