[Cashtab][Alias] Update the alias available input message
Summary:
Per UAT feedback, color the alias length/price message as eCashBlue to be consistent with the purple 'alias is taken' error message for the alias input field. The message itself is also updated to be more obvious this is available.

{F7133764}

Existance of pending alias warning is still handled in the subsequent confirmation modal as usual.

Test Plan:
npm test
npm start
Input an invalid alias and ensure the input validation error is triggered
Input a valid registered alias and ensure the existing alias is taken error is displayed
Input a valid unregistered alias and ensure the `This (length) byte alias is available, (price) XEC to register.` is displayed in eCashBlue.
Make sure there is at least one pending alias for this wallet and then click Register on the same pending alias and ensure no regression to the pending warning on the subsequent confirmation modal.

Reviewers: bytesofman, #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14919