[Chronik] Improve error message when a UTXO couldn't be removed from the mempool
Summary:
There are some issues with mempool UTXOs when blocks are being reorged by Avalanche currently, but the error message doesn't tell us if it was because of a removed or added tx. This makes it
so we know which tx caused the issue, and whether it was because of removal or addition of txs.

Test Plan: `ninja check-crates`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D14935