[Cashtab] Remove remote code from extension
Summary: 2.0.1 did not pass review bc chrome store does not allow remote code. The remote code, sideshift, is not actually used in the extension (for this reason). However it is still in `index.html` and should not be for the extension.

Test Plan:
`npm run extension`, confirm that "SIDESHIFT" does not appear anywhere in `extension/dist/index.html`

`npm run build`, confirm 'SIDESHIFT" does appear in `build/index.html`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14954