[avalanche] Remote proofs: only account for the stakes
Summary:
The current heuristics requires most of the nodes and most of the stakes to be present/absent to have a conclusive remote proofs status.
This diff removes the reliance on the node count, which is not the security model of avalanche and as such adds only complexity and bias.

Depends on D14957.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D14958