[CI] Run the chronik client  integrations tests on diffs
Summary:
Run the integration test each time a change is made to chronik or the chronik-client ts library.
Note that on CI we don't have a python executable, so be explicit and use python3.

I noticed some nyc generated files appearing in my worktree so I added a .gitignore entry for them.

Test Plan:
  ./contrib/teamcity/build-configurations.py chronik-client-integration-tests

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14971