[Cashtab] Extend alias validation to include any input
Summary: Function was created to validate user input, which we can safely assume is always a string. However, we will also want to use this function to validate aliases for creating alias registration txs. In this case, the dev "should" always pass a string. But we can't really be sure without asserting it.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15022