[explorer] Fix the build
Summary:
The build is currently broken because we require a package that depends on rust > 1.70, and we still use 1.66.
This diff updates the version and use 1.72 instead to match our current CI version, and bookworm instead of bullseye to get rid of a CVE in ncurse.

Test Plan:
  cd web/explorer
  docker build -t explorer .
  docker run -p 3035:3035 -t explorer
Check it runs on localhost:3035

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15026