[ecashaddrjs] Copy all files to support prepublish command in CI
Summary:
Support autodeploy for ecashaddrjs

ecashaddrjs, unlike ecash-coinselect, must build before deployment -- the browser version is built with webpack, while the nodejs version just uses the existing JS. This is because browsers do not tend to support the Buffer type which is used in the library.

The `npm ci` command was failing in docker build bc it requires the directory structure to be intact due to the `prepublish` command. I didn't know this but apparently `npm ci` will also run `npm prepublish`.

I was able to repeat the same error msg thrown by teamcity with this testplan, which is corrected with the changes in this diff.

Test Plan: `docker build -t ecashaddrjs_local` succeeds

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15029