[Cashtab] Update param parsing function to support opreturn param
Summary:
Cashtab currently only supports the `amount` param.

Update `parseAddressForParams` function to also support the `opreturn` param.

Ref proof of concept D15032

Note: this diff does not yet implement the ability to create and broadcast txs using this param, although it does update a function currently implemented on the Send.js screen. This diff does not break existing `amount` functionality. For now, if a valid `opreturn` param is in a query string, it is ignored.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15075