[Cashtab] Update webpack to support cjs
Summary:
Known issue with create-react-app. As more dependencies use these files, will need this support.

Main issue now is that the latest chronik-client won't work, since it uses the latest version of axios, which uses cjs files.

Cashtab will work with the latest version of chronik-client with this webpack config.

I can wait and add them together so there is a test plan, but imo it makes sense to keep this as an isolated change.

This is more or less a backport of (still unmerged) https://github.com/facebook/create-react-app/pull/12021

Cashtab was initially a create-react-app app, but has been ejected and customized to support crypto dependencies.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15077