[Cashtab] Add missing dep to dev deps
Summary:
Cashtab was originally a create-react-app project. It has since been "ejected" -- meaning all of the hidden logic built into create-react-app has been converted into specific config files, such as webpack config. These files have been maintained and updated over time.

So, Cashtab is not really a create-react-app project anymore. However, since create-react-app uses ppular react and webpack dependencies, many of these persist.

Recently this warning appeared when running `npm run build`

```
Creating an optimized production build...
One of your dependencies, babel-preset-react-app, is importing the
"@babel/plugin-proposal-private-property-in-object" package without
declaring it in its dependencies. This is currently working because
"@babel/plugin-proposal-private-property-in-object" is already in your
node_modules folder for unrelated reasons, but it may break at any time.

babel-preset-react-app is part of the create-react-app project, which
is not maintianed anymore. It is thus unlikely that this bug will
ever be fixed. Add "@babel/plugin-proposal-private-property-in-object" to
your devDependencies to work around this error. This will make this message
go away.
```

Add @babel/plugin-proposal-private-property-in-object to dev dependencies so that it is not lost in the event of changes to the other dependency coincidentally requiring it (and thus allowing the builds to pass).

Test Plan: `npm run build`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15132