[Cashtab] Stop using a function that isnt really used in SignVerify component
Summary:
We don't support params in this screen so we should not be using the `parseAddressForParams` function. We are just passing user input there to get it back in an object at the `address` key.

Stop doing this.

Ref D15134 for the bigger picture here -- we need more granular info in the `parseAddressForParams` function to support additional bip21 parameters like `opreturn`

Test Plan:
`npm test`

`npm start`, navigate to sign verify screen, paste a valid ecash: address in verify box and note no validation, paste an invalid address and note error msg, paste an etoken: addr and note error msg

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15135