[bip21] Support multiple outputs
Summary:
T3401

It would be useful if there were a standard for creating multi-output transactions with a URI string.

This is currently supported in an unspec'd way by both Cashtab and ElectrumABC with the use of CSV strings.

With webapps looking to create these transactions programmatically, we need a settled spec for how this info is communicated between a webapp and a wallet, or a QR code and a wallet.

One option is to continue the use of CSV strings. However, this would not allow us to also include the `opreturn` param without a new spec, and it seems dumb to have two specs.

Test Plan: Review and suggest improvements

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: Fabien, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15160