[Cashtab] Add chaintip blockheight to wallet context
Summary:
We need the chaintip blockheight to determine which coinbase utxos are not yet mature.

Add this info to wallet context.

If the info is unavailable, Cashtab will ignore all coinbase utxos. So, if there is some error in fetching it, log it to the console but do not lock the UI.

Test Plan:
Add `console.log(`tipHeight`, tipHeight)` between lines 241 and 242 in `useWallet.js`
`npm start`, check that `tipHeight` is printed to the console.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15180