[Cashtab] Add more wallet helper functions
Summary:
Add standardized `toXec` and `toSatoshis` functions. Cashtab has iterated through too many of these functions. They should be simple, tested, and live in one place.

For now -- just use the BigNumber library, as this can't be deprecated for some time.

This diff does not implement the functions. This is part of a series of diffs working to standardize balance headers in Cashtab and remove functions from `utils`.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15202