[Cashtab] Implement unified balance header
Summary:
Combine BalanceHeader and BalanceHeaderFiat into a cleaner, unit-tested component. Implement in App.js so it appears in the same position for all screens.

Note: This means the Settings screen how has the BalanceHeader at the top which, imo, is an improvment.

Features of new BalanceHeader
- Hidden balance now controlled by css and not component replacement (faster)
- Support for different user locale number formatting
- Decimal places are now always constant (i.e. 2 decimal places for fiat and XEC, not sometimes 1)
- Unit tested component
- Unit tests for App.js

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15205