[electrum] escape special chars in URI for op_return message
Summary:
As per title

Depends on D15236

Test Plan:
`python test_runner.py`

In the "receive" tab, enter weird stuff in the OP_RETURN field with the "Raw hex script" option unchecked. Click the "Copy URI", paste it in a notebook and check that special chars are escaped and URI-safe.

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D15231