[electrum] add tests for BIP21 parsing and creation
Summary:
Add an op_return test case for TestParseURI and  add a suite for `create_URI`.
The new test suite shows that currently special chars are not escaped for the op_return paramater. This will be fixed in a separate diff.

Depends on D15235

Test Plan: `python test_runner.py`

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D15236