[Cashtab] Remove unreachable loadStoredWallet migration helper fn
Summary:
With D15256 landed, we can no longer pass `isValidStoredWallet` with the type of wallet that would benefit from the limited migration implemented in the confusingly named `loadStoredWallet` function

Delete this function. If Cashtab loads with an invalid wallet, it will populate wallet state from chronik and not by modifying params in the local storage object.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15243