[electrum] fix cashaddr prefix for fulcrum donation address
Summary:
We no longer support the "bitcoincash:" prefix, so if a user clicks on "Help > Donate to server" while connected to a  server that advertises the default donation address, it causes an error. Fix this by converting the address explicitely to an "ecash:" address.

Fix also two prefixes in comments.

Depends on D15250

Test Plan: Connect to a server that does not specify a custom donation address (`electrum.bitcoinabc.org`) and click the "Help > Donate to server" menu action. Check that it does not raise an error message.

Reviewers: #bitcoin_abc, bytesofman, Fabien

Reviewed By: #bitcoin_abc, bytesofman, Fabien

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D15251