[Cashtab] Stop supporting no-API loading of legacy wallet formats
Summary:
If Cashtab loads a wallet from local storage and finds it is valid, it will immediately set this as the active wallet.

If it is not valid, it will wait to get the valid wallet from the API. With chronik, this now happens immediately on load.

We do not get any performance benefit from non-API migration of very old wallets when we can simply wait 10 ms. Stop supporting this.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15256