[Cashtab] Add used dependency to dev dependencies
Summary:
We use prop-types to throw warnings in dev environment if components have props of the wrong type

This used to be bundled with React by default. It isn't anymore. Import is still working as it is a dependency of other things like eslint. But it should be specified as a dev dependency of Cashtab.

Test Plan: `npm test`, `npm run build`

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15261