[refactor] Make signals optional in mempool and chainman
Summary:
This is done in preparation for the next two commits, where the
CMainSignals are de-globalized.

This avoids adding new constructor arguments to the ChainstateManager
and CTxMemPool classes over the next two commits.

This could also allow future tests that are only interested in the
internal behaviour of the classes to forgo instantiating the signals.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/28960 | core#28960]]
https://github.com/bitcoin/bitcoin/pull/28960/commits/3fba3d5deec6d7bae33823b8da7682f9b03d9deb

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20501