[e.cash] modify CSP headers for dev mode
Summary:
The site was not running right in dev mode with the newly enforced CSP headers. Adding a conditional value for dev mode
that should fix this

Test Plan:
to see the issue: npm run dev on master branch
Should see basically all of the content blocked from headers

then try npm run dev on this branch. Site should load as expected
Can go to the networks tab and check localhost response headers to see the CSP values

then preview with the bot and check the headers to see 'self' is no longer there

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D15266