[Cashtab] Display Cashtab webapp version on settings page if available
Summary:
Display Cashtab version on settings page.

This way we can tell what version a user is getting on their device if they are experiencing an issue, and confirm whether or not the issue is related to having a cached version.

Note: we want the version from package.json, and not package-lock.json, as the version number in package-lock.json only changes to match package.json when the dependencies change. So, the version in package-lock.json is the last version where the dependencies changed, not necessarily the most recent version.

Will prob want some kind of a lint strategy for this soon, but will see how it works manually. The type of version bump for now is the kind of thing that should be reviewed. This is an improvement over no version mgmt.

Test Plan:
`npm start`

Navigate to Configure.js and observe the version at the bottom of page
{F7274581}

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15275