Clarify the transaction validation result
Summary:
The TX_CONFLICT validation result is used in several places that are not related to tx conflict. This diff creates 2 new codes to avoid confusion, and actually use the TX_CONFLICT value when there is a conflict (instead of using a policy error value).

This will make it easier to spot the conflicting transaction and take action like storing them and polling via avalanche.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15283